This is several changes to `ServeMux` in one PR:
- The zero `ServeMux` is now usable, `NewServeMux` doesn't need to be called.
- A series of documentation changes.
- Changes to make `ServeMux.ServeDNS` more readable.
- Performance improvements, and simplification, in `ServeMux.match`. Note this is complicated somewhat by ^1.

```
name                   old time/op    new time/op    delta
MuxMatch/lowercase-12     108ns ± 3%      68ns ± 2%   -36.94%  (p=0.000 n=10+10)
MuxMatch/uppercase-12     111ns ± 3%     120ns ± 1%    +7.34%  (p=0.000 n=9+10)

name                   old alloc/op   new alloc/op   delta
MuxMatch/lowercase-12     32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
MuxMatch/uppercase-12     32.0B ± 0%     32.0B ± 0%      ~     (all equal)

name                   old allocs/op  new allocs/op  delta
MuxMatch/lowercase-12      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
MuxMatch/uppercase-12      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
```

^1: `strings.ToLower` contains one extra allocation in the uppercase case that will be fixed by [CL 137575](https://go-review.googlesource.com/c/go/+/137575). Until that lands, I've reimplemented an ASCII-only `ToLower` that avoids the allocation. This complicates the diff and only improves the case where `q` (the domain name) has uppercase characters. @miekg If this is uncommon, I'd be very happy to revert that part of this PR, but it will temporarily worsen the uppercase performance.