@tmthrgd 2018-09-26
[CL 137575](https://go-review.googlesource.com/c/go/+/137575) has now been merged as golang/go@14e7f174c1, meaning it will appear in go1.12. That should be released at the start of February next year ([see here](https://pocketgophers.com/go-release-timeline/#go1.12)).